add docblocks + internalize some types
