+++ Sources/Leaf/Parser/Byte+Leaf.swift
@@ -1,7 +1,7 @@
 import Bits
 
 /// Leaf specific byte helpers
-extension Byte {
+internal extension Byte {
     var isAllowedInIdentifier: Bool {
         return isAlphanumeric || self == .hyphen || self == .underscore || self == .colon || self == .period
     }
