+++ Sources/Leaf/Renderer/LeafRenderer.swift
@@ -1,10 +1,3 @@
-import Async
-import Bits
-import Dispatch
-import Foundation
-import Service
-import TemplateKit
-
 /// Renders Leaf templates using the Leaf parser and serializer.
 public final class LeafRenderer: ViewRenderer, TemplateRenderer, Service {
     /// See TemplateRenderer.parser
